// admin/src/api/generated/models/ApplicationConfigFileResponse.ts
export interface ApplicationConfigFileResponse {
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    applicationId: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    bindingId: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    content?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    currentDigest?: string | null;
    /**
     *
     * @type {number}
     * @memberof ApplicationConfigFileResponse
     */
    currentVersion: number;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    deletedAt?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    delivery: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    deployPath?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    description: string;
    /**
     *
     * @type {boolean}
     * @memberof ApplicationConfigFileResponse
     */
    editable: boolean;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    format: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    label: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    language: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    path: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    recommendedChanges: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    role: string;
    /**
     *
     * @type {boolean}
     * @memberof ApplicationConfigFileResponse
     */
    sensitive: boolean;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    status: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    templateSourceDigest?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigFileResponse
     */
    updatedAt: string;
    /**
     *
     * @type {number}
     * @memberof ApplicationConfigFileResponse
     */
    version: number;
}