// admin/src/api/generated/models/ApplicationConfigVersionResponse.ts
export interface ApplicationConfigVersionResponse {
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    applicationConfigFileId: string;
    /**
     *
     * @type {number}
     * @memberof ApplicationConfigVersionResponse
     */
    configVersion: number;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    createdAt: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    createdBy?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    digest?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    source: string;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    sourceTemplateDigest?: string | null;
    /**
     *
     * @type {string}
     * @memberof ApplicationConfigVersionResponse
     */
    sourceVersionId?: string | null;
}