// admin/src/api/generated/models/ConfigDiagnostic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * deploy-go-api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.2.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface ConfigDiagnostic
 */
export interface ConfigDiagnostic {
    /**
     *
     * @type {string}
     * @memberof ConfigDiagnostic
     */
    code: string;
    /**
     *
     * @type {number}
     * @memberof ConfigDiagnostic
     */
    column: number;
    /**
     *
     * @type {number}
     * @memberof ConfigDiagnostic
     */
    line: number;
    /**
     *
     * @type {string}
     * @memberof ConfigDiagnostic
     */
    message: string;
    /**
     *
     * @type {string}
     * @memberof ConfigDiagnostic
     */
    path: string;
}

/**
 * Check if a given object implements the ConfigDiagnostic interface.
 */
export function instanceOfConfigDiagnostic(value: object): value is ConfigDiagnostic {
    if (!('code' in value) || value['code'] === undefined) return false;
    if (!('column' in value) || value['column'] === undefined) return false;
    if (!('line' in value) || value['line'] === undefined) return false;
    if (!('message' in value) || value['message'] === undefined) return false;
    if (!('path' in value) || value['path'] === undefined) return false;
    return true;
}

export function ConfigDiagnosticFromJSON(json: any): ConfigDiagnostic {
    return ConfigDiagnosticFromJSONTyped(json, false);
}

export function ConfigDiagnosticFromJSONTyped(json: any, ignoreDiscriminator: boolean): ConfigDiagnostic {
    if (json == null) {
        return json;
    }
    return {

        'code': json['code'],
        'column': json['column'],
        'line': json['line'],
        'message': json['message'],
        'path': json['path'],
    };
}

export function ConfigDiagnosticToJSON(json: any): ConfigDiagnostic {
    return ConfigDiagnosticToJSONTyped(json, false);
}

export function ConfigDiagnosticToJSONTyped(value?: ConfigDiagnostic | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'code': value['code'],
        'column': value['column'],
        'line': value['line'],
        'message': value['message'],
        'path': value['path'],
    };
}